agent. Split agent.go refactor
Long file contents are split into hbuiltin, hevent and htask .go files.
